notmuch-reply: Remove bcc header from notmuch reply.
Notmuch reply should not be Bcc'ing the sender by default.  This is
not the appropriate way to save copies of sent mail (which should
probably be handled by an Fcc header) and it doesn't give the user the
option to not be bcc'd.  This is really something that should be
handled by the reader UI.  For instance, emacs message-mode can easily
be configured to add Bcc's if the user wishes.

This patch, in combination with e.g. the solution in mail id:878wbj4nfq.fsf@servo.finestructure.net allows seamless FCC handling.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>