Restructure notmuch-hello-reflect
This function was "not lispy" according to a comment by dme, and it
also relied on a helper function that used 'cl code. Do away with all
that and provide a hopefully lispy solution that relies on elisp only
to "reflect" a (possibly non-square) matrix along it's diagonal.

Remove now unused notmuch-hello-reflect-generate-row function.

Remove unused notmuch-hello-roundup function (which did exactly the
same as (ceiling divident divisor) anyway).

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>